fix(rendering): correct Go API calls and type casts
- Replace IntN with Intn (correct Go math/rand method)
- Add float32 casts for ebiten ColorScale.Scale parameters
- Rename clampFloat to clampFloatLighting to avoid redeclaration
- Update loop.log and test output

Fixes multiple compilation errors in rendering packages.

+++ pkg/rendering/genre_overlay.go
@@ -38,7 +38,7 @@ func (go_ *GenreOverlay) ApplyOverlay(img *ebiten.Image) *ebiten.Image {
 	overlay := go_.getOverlayParams()
 
 	// Apply color tinting based on genre
-	op.ColorScale.Scale(overlay.tintR, overlay.tintG, overlay.tintB, 1.0)
+	op.ColorScale.Scale(float32(overlay.tintR), float32(overlay.tintG), float32(overlay.tintB), 1.0)
 
 	result.DrawImage(img, op)
